pyproject.toml
[project]
name = "sync-daddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.5",
    "types-pyyaml>=6.0.12.20250915",
]
